+++ src/params.h
@@ -105,6 +105,7 @@ struct compression_params {
     std::string input_filename_2;
     std::string note;
     std::string assay;
+    std::string assay_confidence;
     uint32_t sequence_crc[2];
     uint32_t quality_crc[2];
     uint32_t id_crc[2];